rename: graph & privy type files
