--- types/data.ts
@@ -1,4 +1,0 @@
-export type DataPoint = {
-    value: number;
-    label: string;
-};
