+++ types/graph.ts
@@ -1,0 +1,4 @@
+export type DataPoint = {
+  value: number;
+  label: string;
+};
